esbd3_part_1/atividade/first_stage/requirements.txt
numpy
pandas
scikit-learn
joblib
pyzmq
minio